refactor(web): split runs data helpers [OPE-348]
